hal/gles: command module
