+++ wgpu-hal/src/gles/mod.rs
@@ -4,6 +4,7 @@
 mod egl;
 
 mod adapter;
+mod command;
 mod conv;
 mod device;
 
@@ -12,12 +13,11 @@ use self::egl::{Instance, Surface};
 
 use glow::HasContext;
 
-use std::{ops::Range, sync::Arc};
+use std::sync::Arc;
 
 #[derive(Clone)]
 pub struct Api;
 pub struct Context;
-pub struct Encoder;
 #[derive(Debug)]
 pub struct Resource;
 
@@ -30,8 +30,8 @@ impl crate::Api for Api {
     type Device = Device;
 
     type Queue = Queue;
-    type CommandEncoder = Encoder;
-    type CommandBuffer = Resource;
+    type CommandEncoder = CommandEncoder;
+    type CommandBuffer = CommandBuffer;
 
     type Buffer = Buffer;
     type Texture = Texture;
@@ -221,10 +221,23 @@ pub struct Sampler {
     raw: glow::Sampler,
 }
 
+#[derive(Debug)]
+enum Command {}
+
+#[derive(Default)]
+pub struct CommandBuffer {
+    commands: Vec<Command>,
+    data: Vec<u8>,
+}
+
+pub struct CommandEncoder {
+    cmd_buffer: CommandBuffer,
+}
+
 impl crate::Queue<Api> for Queue {
     unsafe fn submit(
         &mut self,
-        command_buffers: &[&Resource],
+        command_buffers: &[&CommandBuffer],
         signal_fence: Option<(&mut Resource, crate::FenceValue)>,
     ) -> DeviceResult<()> {
         Ok(())
@@ -236,165 +249,4 @@ impl crate::Queue<Api> for Queue {
     ) -> Result<(), crate::SurfaceError> {
         Ok(())
     }
-}
-
-impl crate::CommandEncoder<Api> for Encoder {
-    unsafe fn begin_encoding(&mut self, label: crate::Label) -> DeviceResult<()> {
-        Ok(())
-    }
-    unsafe fn discard_encoding(&mut self) {}
-    unsafe fn end_encoding(&mut self) -> DeviceResult<Resource> {
-        Ok(Resource)
-    }
-    unsafe fn reset_all<I>(&mut self, command_buffers: I) {}
-
-    unsafe fn transition_buffers<'a, T>(&mut self, barriers: T)
-    where
-        T: Iterator<Item = crate::BufferBarrier<'a, Api>>,
-    {
-    }
-
-    unsafe fn transition_textures<'a, T>(&mut self, barriers: T)
-    where
-        T: Iterator<Item = crate::TextureBarrier<'a, Api>>,
-    {
-    }
-
-    unsafe fn fill_buffer(&mut self, buffer: &Buffer, range: crate::MemoryRange, value: u8) {}
-
-    unsafe fn copy_buffer_to_buffer<T>(&mut self, src: &Buffer, dst: &Buffer, regions: T) {}
-
-    unsafe fn copy_texture_to_texture<T>(
-        &mut self,
-        src: &Texture,
-        src_usage: crate::TextureUse,
-        dst: &Texture,
-        regions: T,
-    ) {
-    }
-
-    unsafe fn copy_buffer_to_texture<T>(&mut self, src: &Buffer, dst: &Texture, regions: T) {}
-
-    unsafe fn copy_texture_to_buffer<T>(
-        &mut self,
-        src: &Texture,
-        src_usage: crate::TextureUse,
-        dst: &Buffer,
-        regions: T,
-    ) {
-    }
-
-    unsafe fn begin_query(&mut self, set: &Resource, index: u32) {}
-    unsafe fn end_query(&mut self, set: &Resource, index: u32) {}
-    unsafe fn write_timestamp(&mut self, set: &Resource, index: u32) {}
-    unsafe fn reset_queries(&mut self, set: &Resource, range: Range<u32>) {}
-    unsafe fn copy_query_results(
-        &mut self,
-        set: &Resource,
-        range: Range<u32>,
-        buffer: &Buffer,
-        offset: wgt::BufferAddress,
-    ) {
-    }
-
-    // render
-
-    unsafe fn begin_render_pass(&mut self, desc: &crate::RenderPassDescriptor<Api>) {}
-    unsafe fn end_render_pass(&mut self) {}
-
-    unsafe fn set_bind_group(
-        &mut self,
-        layout: &Resource,
-        index: u32,
-        group: &Resource,
-        dynamic_offsets: &[wgt::DynamicOffset],
-    ) {
-    }
-    unsafe fn set_push_constants(
-        &mut self,
-        layout: &Resource,
-        stages: wgt::ShaderStage,
-        offset: u32,
-        data: &[u32],
-    ) {
-    }
-
-    unsafe fn insert_debug_marker(&mut self, label: &str) {}
-    unsafe fn begin_debug_marker(&mut self, group_label: &str) {}
-    unsafe fn end_debug_marker(&mut self) {}
-
-    unsafe fn set_render_pipeline(&mut self, pipeline: &Resource) {}
-
-    unsafe fn set_index_buffer<'a>(
-        &mut self,
-        binding: crate::BufferBinding<'a, Api>,
-        format: wgt::IndexFormat,
-    ) {
-    }
-    unsafe fn set_vertex_buffer<'a>(&mut self, index: u32, binding: crate::BufferBinding<'a, Api>) {
-    }
-    unsafe fn set_viewport(&mut self, rect: &crate::Rect<f32>, depth_range: Range<f32>) {}
-    unsafe fn set_scissor_rect(&mut self, rect: &crate::Rect<u32>) {}
-    unsafe fn set_stencil_reference(&mut self, value: u32) {}
-    unsafe fn set_blend_constants(&mut self, color: &wgt::Color) {}
-
-    unsafe fn draw(
-        &mut self,
-        start_vertex: u32,
-        vertex_count: u32,
-        start_instance: u32,
-        instance_count: u32,
-    ) {
-    }
-    unsafe fn draw_indexed(
-        &mut self,
-        start_index: u32,
-        index_count: u32,
-        base_vertex: i32,
-        start_instance: u32,
-        instance_count: u32,
-    ) {
-    }
-    unsafe fn draw_indirect(
-        &mut self,
-        buffer: &Buffer,
-        offset: wgt::BufferAddress,
-        draw_count: u32,
-    ) {
-    }
-    unsafe fn draw_indexed_indirect(
-        &mut self,
-        buffer: &Buffer,
-        offset: wgt::BufferAddress,
-        draw_count: u32,
-    ) {
-    }
-    unsafe fn draw_indirect_count(
-        &mut self,
-        buffer: &Buffer,
-        offset: wgt::BufferAddress,
-        count_buffer: &Buffer,
-        count_offset: wgt::BufferAddress,
-        max_count: u32,
-    ) {
-    }
-    unsafe fn draw_indexed_indirect_count(
-        &mut self,
-        buffer: &Buffer,
-        offset: wgt::BufferAddress,
-        count_buffer: &Buffer,
-        count_offset: wgt::BufferAddress,
-        max_count: u32,
-    ) {
-    }
-
-    // compute
-
-    unsafe fn begin_compute_pass(&mut self, desc: &crate::ComputePassDescriptor) {}
-    unsafe fn end_compute_pass(&mut self) {}
-
-    unsafe fn set_compute_pipeline(&mut self, pipeline: &Resource) {}
-
-    unsafe fn dispatch(&mut self, count: [u32; 3]) {}
-    unsafe fn dispatch_indirect(&mut self, buffer: &Buffer, offset: wgt::BufferAddress) {}
 }
